% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary-distr.r
\name{mg_to_ptm}
\alias{mg_to_ptm}
\alias{am_to_ptm}
\title{Mobility Graph to Probability Transition Matrix}
\usage{
mg_to_ptm(am)
}
\arguments{
\item{am}{the adjacency matrix.}
}
\description{
Mobility Graph to Probability Transition Matrix
}
