% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.r
\name{mobility_asymmetry}
\alias{mobility_asymmetry}
\title{Mobility Asymmetry}
\usage{
mobility_asymmetry(mg, prop = TRUE)
}
\arguments{
\item{mg}{the mobility graph.}

\item{prop}{should the mobility asymmetry be reported as a proportion?
Default TRUE.}
}
\description{
Mobility Asymmetry
}
