% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary-distr.r
\name{stationary_distr}
\alias{stationary_distr}
\title{Find the Stationary Distribution of a Markov Transition Matrix}
\usage{
stationary_distr(P)
}
\arguments{
\item{P}{a Markov transition matrix. Rows correspond to the "from"
vertex, columns correspond to the "to" vertex.}
}
\description{
Find the Stationary Distribution of a Markov Transition Matrix
}
