% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary-distr.r
\name{stationary_distr}
\alias{stationary_distr}
\title{Find the Stationary Distribution of a Markov Transition Matrix}
\usage{
stationary_distr(P)
}
\arguments{
\item{P}{a Markov transition matrix.}

\item{.eigen}{the function to calculate the eigen vectors with. By
default, when the matrix is small the full eigenvalue decomposition is
peformed. Whe the matrix is large, only the leading 20 eigen vectors are
calculated.}
}
\description{
Find the Stationary Distribution of a Markov Transition Matrix
}
\details{
Note that this
}
